Created separate screens - alarm list, edit alarm and ring alarm
