+++ pubspec.yaml
@@ -50,6 +50,7 @@ dependencies:
   flutter_local_notifications: ^17.1.2
   shared_preferences: ^2.2.3
   pedometer: ^4.0.1
+  bloc: ^8.1.4
 
 dev_dependencies:
   flutter_test:
